+++ tests/test_io.py
@@ -29,6 +29,13 @@ def test_yaml_parser_types_handling(data):
     assert isinstance(item, str)
 
 
+def test_import_from_yaml_args():
+    with pytest.raises(TypeError):
+        import_from_yaml()
+    with pytest.raises(TypeError):
+        import_from_yaml('A', filepath='B')
+
+
 class TestImportFromYaml:
     def test_import_example_from_tests(self, example_from_tests):
         assert isinstance(example_from_tests, Statechart)